Refactor supported group extension tests into their own file.
